+++ packages/bench/sql/schema.sql
@@ -1,30 +1,32 @@
 -- Bench fixture schema.
 -- Single bench table covering text / int / jsonb encrypted columns plus the
--- three canonical EQL functional indexes: hmac_256 (hash), bloom_filter (GIN),
--- ste_vec (GIN).
+-- three canonical EQL v3 functional indexes over the column-side index terms:
+-- eq_term (hash), match_term (GIN bloom), ste_vec (GIN).
 --
--- We deliberately do NOT create the `eql_v2.encrypted_operator_class` btree
--- indexes that ore-benches uses. Encrypted composites for full-feature columns
--- (equality + match + ORE) blow past the 2704-byte btree page-size limit, and
--- those indexes don't exist on Supabase anyway — the bench's whole job is to
--- validate that the functional-index path works.
+-- Mirrors `src/drizzle/setup.ts`: the columns are concrete `eql_v3_*` Postgres
+-- domains (see `types.TextSearch` / `types.IntegerOrd` / `types.Json`). Install
+-- the domains and index-term functions first with `stash eql install --eql-version 3`.
+--
+-- We deliberately do NOT create btree operator-class indexes: encrypted terms for
+-- full-feature columns blow past the 2704-byte btree page-size limit, and the
+-- bench's whole job is to validate that the functional-index path works.
 
 DROP TABLE IF EXISTS bench;
 
 CREATE TABLE bench (
     id        SERIAL PRIMARY KEY,
-    enc_text  eql_v2_encrypted NOT NULL,
-    enc_int   eql_v2_encrypted NOT NULL,
-    enc_jsonb eql_v2_encrypted NOT NULL
+    enc_text  public.eql_v3_text_search NOT NULL,
+    enc_int   public.eql_v3_integer_ord NOT NULL,
+    enc_jsonb public.eql_v3_json_search NOT NULL
 );
 
 CREATE INDEX bench_text_hmac_idx
-    ON bench USING hash (eql_v2.hmac_256(enc_text));
+    ON bench USING hash (eql_v3.eq_term(enc_text));
 
 CREATE INDEX bench_text_bloom_idx
-    ON bench USING gin  (eql_v2.bloom_filter(enc_text));
+    ON bench USING gin  (eql_v3.match_term(enc_text));
 
 CREATE INDEX bench_jsonb_stevec_idx
-    ON bench USING gin  (eql_v2.ste_vec(enc_jsonb));
+    ON bench USING gin  (eql_v3.ste_vec(enc_jsonb));
 
 ANALYZE bench;
