+++ packages/stack-drizzle/__tests__/schema-extraction.test.ts
@@ -1,10 +1,10 @@
 import { encryptedTable, types as v3Types } from '@cipherstash/stack/eql/v3'
 import { integer, pgTable } from 'drizzle-orm/pg-core'
 import { describe, expect, it } from 'vitest'
-import { extractEncryptionSchemaV3 } from '../../src/v3/schema-extraction'
-import { types } from '../../src/v3/types'
+import { extractEncryptionSchema } from '../src/schema-extraction'
+import { types } from '../src/types'
 
-describe('extractEncryptionSchemaV3', () => {
+describe('extractEncryptionSchema', () => {
   it('rebuilds an equivalent eql/v3 encryptedTable from every drizzle v3 factory', () => {
     const drizzleColumns = Object.fromEntries(
       Object.entries(types).map(([factoryName, factory]) => [
@@ -19,7 +19,7 @@ describe('extractEncryptionSchemaV3', () => {
       ]),
     )
 
-    const extracted = extractEncryptionSchemaV3(
+    const extracted = extractEncryptionSchema(
       pgTable('users', drizzleColumns as never),
     )
     const authored = encryptedTable('users', authoredColumns as never)
@@ -34,7 +34,7 @@ describe('extractEncryptionSchemaV3', () => {
       age: types.IntegerOrd('age'),
     })
 
-    const extracted = extractEncryptionSchemaV3(users)
+    const extracted = extractEncryptionSchema(users)
     const authored = encryptedTable('users', {
       email: v3Types.TextSearch('email'),
       age: v3Types.IntegerOrd('age'),
@@ -51,8 +51,8 @@ describe('extractEncryptionSchemaV3', () => {
       email: types.IntegerOrd('email'),
     })
 
-    const accountsSchema = extractEncryptionSchemaV3(accounts)
-    const metricsSchema = extractEncryptionSchemaV3(metrics)
+    const accountsSchema = extractEncryptionSchema(accounts)
+    const metricsSchema = extractEncryptionSchema(metrics)
 
     expect(accountsSchema.build()).toStrictEqual(
       encryptedTable('accounts', {
@@ -72,7 +72,7 @@ describe('extractEncryptionSchemaV3', () => {
       emailAddress: types.TextEq('email_address'),
     })
 
-    expect(extractEncryptionSchemaV3(users).build()).toStrictEqual(
+    expect(extractEncryptionSchema(users).build()).toStrictEqual(
       encryptedTable('users', {
         createdOn: v3Types.Date('created_on'),
         emailAddress: v3Types.TextEq('email_address'),
@@ -89,7 +89,7 @@ describe('extractEncryptionSchemaV3', () => {
       },
     }
 
-    expect(extractEncryptionSchemaV3(table as never).build()).toStrictEqual(
+    expect(extractEncryptionSchema(table as never).build()).toStrictEqual(
       encryptedTable('users', {
         createdOn: v3Types.Date('created_on'),
       }).build(),
@@ -98,12 +98,12 @@ describe('extractEncryptionSchemaV3', () => {
 
   it('throws when the table has no encrypted v3 columns', () => {
     const plain = pgTable('plain', { id: integer() })
-    expect(() => extractEncryptionSchemaV3(plain)).toThrow(/no encrypted v3/i)
+    expect(() => extractEncryptionSchema(plain)).toThrow(/no encrypted v3/i)
   })
 
   it('throws the table-name error before checking for encrypted columns', () => {
     expect(() =>
-      extractEncryptionSchemaV3({
+      extractEncryptionSchema({
         secret: { getSQLType: () => 'public.eql_v3_text_eq', name: 'secret' },
       } as never),
     ).toThrow('Unable to read table name from Drizzle table.')
