+++ packages/stack-drizzle/__tests__/types.test-d.ts
@@ -1,7 +1,7 @@
 import type { types as v3Types } from '@cipherstash/stack/eql/v3'
 import type { Encrypted } from '@cipherstash/stack/types'
 import { describe, expectTypeOf, it } from 'vitest'
-import { types } from '../../src/v3/types'
+import { types } from '../src/types'
 
 describe('v3 drizzle types - type-level', () => {
   it('exposes exactly the same factory keys as @/eql/v3 types', () => {
