+++ packages/stack-drizzle/package.json
@@ -43,24 +43,7 @@
 				"default": "./dist/index.cjs"
 			}
 		},
-		"./v3": {
-			"import": {
-				"types": "./dist/v3/index.d.ts",
-				"default": "./dist/v3/index.js"
-			},
-			"require": {
-				"types": "./dist/v3/index.d.cts",
-				"default": "./dist/v3/index.cjs"
-			}
-		},
 		"./package.json": "./package.json"
-	},
-	"typesVersions": {
-		"*": {
-			"v3": [
-				"./dist/v3/index.d.ts"
-			]
-		}
 	},
 	"scripts": {
 		"prebuild": "node -e \"require('node:fs').rmSync('dist',{recursive:true,force:true})\"",
