--- packages/stack-drizzle/src/v3/index.ts
@@ -1,8 +1,0 @@
-export { EqlV3CodecError, v3FromDriver, v3ToDriver } from './codec.js'
-export { getEqlV3Column, isEqlV3Column, makeEqlV3Column } from './column.js'
-export {
-  createEncryptionOperatorsV3,
-  EncryptionOperatorError,
-} from './operators.js'
-export { extractEncryptionSchemaV3 } from './schema-extraction.js'
-export { types } from './types.js'
